Move get_available_locales() to i18n module
