+++ cddagl/i18n.py
@@ -1,7 +1,7 @@
+import builtins
 import gettext
 import logging
-import builtins
-
+import os
 
 logger = logging.getLogger(__name__)
 
@@ -30,6 +30,19 @@ def proxy_ngettext(*args, **kwargs):
     #if 'ngettext' not in builtins.__dict__:
     #    load_gettext_no_locale()
     return builtins.__dict__['ngettext'](*args, **kwargs)
+
+
+def get_available_locales(locale_dir):
+    """Return a list of available locales in the specified directory."""
+    available_locales = []
+    if os.path.isdir(locale_dir):
+        entries = os.scandir(locale_dir)
+        for entry in entries:
+            if entry.is_dir():
+                available_locales.append(entry.name)
+
+    available_locales.sort(key=lambda x: 0 if x == 'en' else 1)
+    return available_locales
 
 
 def load_gettext_no_locale():
